common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Common utils for the best movie theater"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary==2.9.10",
    "redis==5.2.1",
    "elasticsearch[async]==8.13.2",
    "elasticsearch==8.13.2",
    "fastapi==0.115.11",
    "redis==5.2.1",
    "email-validator==2.2.0",
    "PyJWT==2.10.1",
    "pycryptodomex==3.21.0",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "fastapi==0.115.11",
    "fastapi[standard]==0.115.11",
    "sqlalchemy==2.0.38",
    "python-dateutil==2.9.0",
    "backoff==2.2.1",
    "asyncpg==0.30.0",
    "passlib==1.7.4",
    "Faker==37.1.0",
    "pymongo==4.12.0",
    "aiohttp==3.11.16",
    "aiohappyeyeballs==2.6.1",
    "attrs==25.3.0",
    "pika==1.3.2",
    "httpx==0.28.1"
]
